[normalize][strategies] Move CS-specific part of normalization code to strategies.
